Enhanced the indexer2indexer mapping, for when an indexer has provides id's to other indexers.
The indexer_mapping table in sickbeard.db, is used to keep track of indexer 2 indexer mappings. This was done using a name search. For ex. thetvdb is used to add lost, then the helpers.mapIndexersToShow() tries to look for t['lost'], if it gets one result, the mapping table was updated.

In the new situation, an attribute can be configured for an indexer in indexer_config.py, which other indexers can use to keep track of that indexer in. Now the mapIndexersToShow() function attempts to map based on the id and if that doesn't work tries the name search.