Fixed manual searching for scene named anime shows.
Somehow results = curProvider.cache.update_cache_manual_search(searchResults[curEp]) was always running with curEp=-1. As it should run the function with the episode number as a param.

With this fix it was getting all results for show Fairy Tails S2, for providers nyaatorrent and tokyotosho. As before it only found results from nzb providers.